Notes
The code in timeseries_material_impact_public.m was used to carry out the calculations described in the manuscript, using some utilities that are also posted here, including hypothetical extraction and capital endogenization. The code also includes saving results in a large .mat file and reporting some results to Excel. The DisplayTimeSeries contains most of the figures and saving of data behind figures to Excel.